% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{get_file}
\alias{get_file}
\title{Fetch file paths to sage images}
\usage{
get_file(where, what, index = NULL)
}
\arguments{
\item{where}{where in the document}

\item{what}{type of object}

\item{index}{index of the object, Default: NULL}
}
\value{
character, path to image
}
\description{
Fetch paths to sage images for using in slides and documents.
}
\details{
List of images to select from
\itemize{
\item Title Page
\itemize{
\item title-image
\item title-logo
}
\item Dividers
\itemize{
\item divider-image-1
\item divider-image-2
\item divider-image-3
\item divider-image-4
}
\item Side Content
\itemize{
\item side-content-1
\item side-content-2
\item side-content-3
\item side-content-4
\item side-content-5
\item side-content-6
}
\item Final Slide
\itemize{
\item final-image
}
\item Footer Logo
\itemize{
\item footer-logo
}
}
}
\examples{
 get_file('title','image')
 get_file('divider','image',1)
 get_file('divider','image',2)
}
