% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_template.R
\name{init_workspace}
\alias{init_workspace}
\title{Initialize Workspace}
\usage{
init_workspace(path = ".", gh = "")
}
\arguments{
\item{path}{Path where to copy files into}

\item{gh}{url of github repo that you want to associate with the project}
}
\description{
Initialize workspace for a Sanofi project folder
}
