% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{use_report}
\alias{use_report}
\title{Create report template}
\usage{
use_report(report_name, path = ".", ...)
}
\arguments{
\item{report_name}{name of the report}

\item{path}{Path where to copy files into}

\item{\dots}{unused}
}
\description{
Initialize a report folder structure via officedown
}
