% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{use_slides}
\alias{use_slides}
\title{Create slides template}
\usage{
use_slides(deck_name, path = ".", ...)
}
\arguments{
\item{deck_name}{name of the deck}

\item{path}{Path where to copy files into}

\item{\dots}{unused}
}
\description{
Initialize a slide deck template
}
