% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_template.R
\name{init_workspace}
\alias{init_workspace}
\title{Initialize Workspace}
\usage{
init_workspace(path = ".", ...)
}
\arguments{
\item{path}{Path where to copy files into}

\item{\dots}{arguments passed to \link[renv:init]{init}}
}
\description{
Initialize workspace for a Sanofi project folder
}
